Front end/requirements.txt
numpy==2.2.0
Pillow==11.0.0
python-dotenv==1.0.1
Requests==2.32.3
streamlit==1.41.1
